Move NotificationEndpointIT to testingsupport package and new unit test
