+++ src/integrationTest/java/uk/gov/hmcts/reform/pcs/testingsupport/endpoint/TestingSupportNotificationIT.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pcs.notify;
+package uk.gov.hmcts.reform.pcs.testingsupport.endpoint;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
@@ -39,7 +39,7 @@ import static uk.gov.hmcts.reform.pcs.hearings.constants.HearingConstants.SERVIC
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.MOCK)
 @AutoConfigureMockMvc
 @ActiveProfiles("integration")
-public class NotificationEndpointIT extends AbstractPostgresContainerIT {
+public class TestingSupportNotificationIT extends AbstractPostgresContainerIT {
 
     private static final String AUTH_HEADER = "Bearer token";
     private static final String SERVICE_AUTH_HEADER = "ServiceAuthToken";
